backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy>=2.0
pydantic>=2.9
python-multipart>=0.0.12
python-dotenv>=1.0
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
joblib>=1.4
pytest>=8.3
httpx>=0.27
